fix: correct input_value type to match InputSource u16 API
Cell<u8> caused a type mismatch with InputSource::to_vcp_value() and
from_vcp_value() which use u16, preventing the Linux app from compiling.

+++ apps/linux/src/input_row.rs
@@ -13,7 +13,7 @@ mod imp {
         pub check_icon: TemplateChild<Image>,
 
         pub monitor_index: Cell<usize>,
-        pub input_value: Cell<u8>,
+        pub input_value: Cell<u16>,
     }
 
     #[glib::object_subclass]
